+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [1.30.0] - 2021-06-29
+
+### Changed
+
+- Extend `template app` to only output required fields, the flag `--defaulting-enabled`
+can be set to false to disable this.
+
 ## [1.29.2] - 2021-06-17
 
 - In the `template cluster` command, the flag `--control-plane-az` is replacing `--master-az`.
@@ -358,7 +365,8 @@ This release supports rendering for CRs:
 - `AppCatalog`
 - `App`
 
-[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v1.29.2...HEAD
+[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v1.30.0...HEAD
+[1.30.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.29.2...v1.30.0
 [1.29.2]: https://github.com/giantswarm/kubectl-gs/compare/v1.29.1...v1.29.2
 [1.29.1]: https://github.com/giantswarm/kubectl-gs/compare/v1.29.0...v1.29.1
 [1.29.0]: https://github.com/giantswarm/kubectl-gs/compare/v1.28.0...v1.29.0